Close PG connections faster to avoid idle-in-transaction timeout.
When checking out a large table from PostGIS to Spatialite, it can
happen that the connection to the PostGIS database is closed serverside
due to the idle-in-transaction timeout settings.

This is in particular the case when the ogr2ogr copy takes longer than
the idle-in-transaction timeout of the PostGIS database. Since ogr2ogr
uses a different database connection than the versioning plugin, the
connection of the latter is left idle-in-transaction and closed
serverside. This results in an exception when the connection is used in
the code after the ogr2ogr call.